Refactored ConnectorTypes
