--- app/cdap/components/ConnectorTypes/ConnectorTypeCard/types.ts
@@ -1,4 +1,0 @@
-export interface IConnectorTypeCardComponentProps {
-  name: string;
-  image: JSX.Element;
-}
