--- app/cdap/components/ConnectorTypes/WelcomeCard/styles.ts
@@ -1,9 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-export const useWelcomeCardComponentStyles = makeStyles(() => ({
-  welcomeText: {
-    fontSize: '24px',
-    fontWeight: 500,
-    height: '36px',
-  },
-}));
