+++ app/cdap/components/ConnectorTypes/index.tsx
@@ -1,25 +1,22 @@
-import { Box, Paper, Typography } from '@material-ui/core';
+import { Box, Paper, styled, Typography } from '@material-ui/core';
 import { fetchConnectors } from 'components/Connections/Create/reducer';
 import { defaultConnectorTypes } from 'components/WrangleHome/constants/defaultConnectorTypes';
 import React, { useEffect, useState } from 'react';
-import WranglerCard from './ConnectorTypeCard';
-import { GetConnectionIcon } from './iconStore';
+import { GetConnectionIcon, GetIcon } from './iconStore';
 import { useConnectorTypesComponentStyles } from './styles';
-import WelcomeCard from './WelcomeCard';
 import { getCurrentNamespace } from 'services/NamespaceStore';
 import { Link } from 'react-router-dom';
 import { UnderLine } from './iconStore';
 
 const ConnectorTypesComponent = () => {
   const classes = useConnectorTypesComponentStyles();
-
+  const welcomeIcon = GetIcon('welcomeIcon');
   const [connectorTypesList, setConnectorTypesList] = useState([]);
 
   const fetchConnectorTypeDetails = async () => {
     // fetching the list of connector types
     const fetchedConnectorTypes = await fetchConnectors();
     const connectorTypes = [...defaultConnectorTypes, ...fetchedConnectorTypes];
-
     // creating list of connector's name & corresponding icon
     const connectorTypesWithImage = connectorTypes.map((connector) => ({
       name: connector.name,
@@ -35,7 +32,13 @@ const ConnectorTypesComponent = () => {
   return (
     <>
       <Paper variant="outlined" elevation={9} className={classes.dashBoard}>
-        <WelcomeCard />
+        <Box className={classes.welcomeCardContainer}>
+          <Box>{welcomeIcon}</Box>
+          <Box className={classes.welcomeTextContainer}>
+            <Typography className={classes.welcomeText}>Hi David</Typography>
+            <Typography className={classes.welcomeText}>Welcome to Wrangler</Typography>
+          </Box>
+        </Box>
         <Typography className={classes.subTitle}>Start data exploration</Typography>
         <Box className={classes.underLine}>{UnderLine}</Box>
         <Paper elevation={0} className={classes.flexContainer}>
@@ -44,11 +47,14 @@ const ConnectorTypesComponent = () => {
               to={`/ns/${getCurrentNamespace()}/datasets-list/${eachConnectorType.name}`}
               className={classes.linkLine}
             >
-              <WranglerCard
-                key={eachConnectorType.name}
-                name={eachConnectorType.name}
-                image={eachConnectorType.image}
-              />
+              <Box className={classes.cardWrapper}>
+                <Box className={classes.wrangleCardContent}>
+                  {eachConnectorType.image}
+                  <Typography variant="body1" className={classes.connectorTypeName}>
+                    {eachConnectorType.name}
+                  </Typography>
+                </Box>
+              </Box>
             </Link>
           ))}
         </Paper>
